Move searchOrganization into endpoints.ts per CodeRabbit
The API function was inlined in search.ts, violating the project's
convention of centralizing endpoint definitions in endpoints.ts.

- Add SearchResult interface and searchOrganization function to endpoints.ts
- Replace search.ts body with re-exports for backward compatibility
  (SearchResult type, searchOrg alias pointing to searchOrganization)
